+++ app/Http/Controllers/karyawan/Bagian.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers\karyawan;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Model\Karyawan\Bagian as bagians;
+use App\Model\Karyawan\Devisi as devisis;
+use App\Model\Superadmin_ukm\U_jabatan_p as jabatans;
+
 use Session;
 class Bagian extends Controller
 {
@@ -12,24 +15,45 @@ class Bagian extends Controller
 
     private $id_karyawan;
     private $id_perusahaan;
+    private $level_jabatan=['Direksi','Manager','Supervisor','Staf'];
 
-    public function __construct()
-    {
-        $this->middleware(function($req, $next){
-            if(empty(Session::get('id_karyawan')) && empty(Session::get('id_perusahaan_karyawan')))
-            {
-                Session::flush();
-                return redirect('login-karyawan')->with('message_login_fail','Waktu masuk anda berakhir, Silahkan login Ulang...!!');
-            }
-            $this->id_karyawan = Session::get('id_karyawan');
-            $this->id_perusahaan = Session::get('id_perusahaan_karyawan');
-            return $next($req);
-        });
+    public function __construct(){
+      $this->middleware(function($req, $next){
+          if(empty(Session::get('id_karyawan')) && empty(Session::get('id_perusahaan_karyawan')))
+          {
+              Session::flush();
+              return redirect('login-karyawan')->with('message_login_fail','Waktu masuk anda berakhir, Silahkan login Ulang...!!');
+          }
+          $this->id_karyawan = Session::get('id_karyawan');
+          $this->id_perusahaan = Session::get('id_perusahaan_karyawan');
+          return $next($req);
+      });
     }
 
     public function index()
     {
-        return view('user.karyawan.section.Bagian.page_default');
+      $data= [
+          'Bagian'=>bagians::where('id_perusahaan', $this->id_perusahaan)->paginate(9),
+          'jabatan'=>jabatans::where('id_perusahaan', $this->id_perusahaan)->orderBy('level_jabatan')->get(),
+          'level_jabatan' => $this->level_jabatan
+      ];
+      if(empty(Session::get('tab2')) && empty(Session::get('tab3')) && empty(Session::get('tab4'))){
+          Session::flash('tab1','tab1');
+      }
+
+      if(!empty(Session::get('tab2'))){
+          Session::flash('tab2',Session::get('tab2'));
+      }
+
+      if(!empty(Session::get('tab3'))){
+          Session::flash('tab3',Session::get('tab3'));
+      }
+
+      if(!empty(Session::get('tab4'))){
+          Session::flash('tab4',Session::get('tab4'));
+      }
+
+        return view('user.karyawan.section.Bagian.page_default',$data);
     }
 
     public function DataBagian()
