+++ resources/views/user/karyawan/include/footer.blade.php
@@ -2,7 +2,7 @@
 <footer class="main-footer">
     <!-- To the right -->
     <div class="pull-right hidden-xs">
-        Untuk menjadi besar, perlu sistem
+        <font color="#DE3106"><b>Menjadi besar, perlu sistem !! </b></font>
     </div>
     <!-- Default to the left -->
     <strong>Copyright &copy; 2019 <a href="#">aturusaha.com</a></strong> All rights reserved.
