+++ resources/views/user/karyawan/section/ModelBisnis/page_edit.blade.php
@@ -3,7 +3,7 @@
 @section('skin')
 <link rel="stylesheet" href="{{ asset('component/bower_components/select2/dist/css/select2.min.css') }}">
 
-<script src="{{ asset('ckeditor/ckeditor.js') }}"></script>
+<script src="https://cdn.ckeditor.com/4.11.4/basic/ckeditor.js"></script>
 @stop
